+++ app/Helpers/Delete/Models/DeleteQuotationRequest.php
@@ -8,20 +8,20 @@
 
 namespace App\Helpers\Delete\Models;
 
-
 use App\Helpers\Delete\DeleteInterface;
+use App\Helpers\Delete\Traits\ChecksExcludedCleanupContacts;
 use Illuminate\Database\Eloquent\Model;
+use Illuminate\Support\Facades\Log;
 
 /**
  * Class DeleteQuotationRequest
- *
- * Relation: 1-n Emails. Action: dissociate
- * Relation: 1-n Documents. Action: dissociate
  *
  * @package App\Helpers\Delete\Models
  */
 class DeleteQuotationRequest implements DeleteInterface
 {
+    use ChecksExcludedCleanupContacts;
+
     private $errorMessage = [];
     private $quotationRequest;
 
@@ -29,10 +29,53 @@ class DeleteQuotationRequest implements DeleteInterface
      *
      * @param Model $quotationRequest the model to delete
      */
-
     public function __construct(Model $quotationRequest)
     {
         $this->quotationRequest = $quotationRequest;
+    }
+
+    /** If it's called by the cleanup functionality, we land on this function, else on the delete function
+     *
+     * @return array
+     * @throws
+     */
+    public function cleanup()
+    {
+        try {
+            if (! $this->canCleanup()) {
+                return $this->errorMessage;
+            }
+
+            return $this->delete();
+        } catch (\Exception $exception) {
+            Log::error('Fout bij opschonen Kansacties', [
+                'exception' => $exception->getMessage(),
+                'errormessages' => implode(' | ', $this->errorMessage),
+            ]);
+
+            $this->errorMessage[] =
+                "Fout bij opschonen Kansacties. (meld dit bij Econobis support)";
+
+            return $this->errorMessage;
+        }
+    }
+
+    public function canCleanup(): bool
+    {
+        $contactId = $this->quotationRequest
+            ->opportunity
+            ?->intake
+            ?->contact_id;
+
+        if ($this->isContactExcludedFromCleanup($contactId)) {
+            $this->errorMessage[] =
+                "Kansactie {$this->quotationRequest->id} kan niet worden opgeschoond: "
+                . "het gekoppelde contact valt in een uitzonderingsgroep.";
+
+            return false;
+        }
+
+        return true;
     }
 
     /** Main method for deleting this model and all it's relations
@@ -42,32 +85,26 @@ class DeleteQuotationRequest implements DeleteInterface
      */
     public function delete()
     {
-        $this->canDelete();
+        if (! $this->canDelete()) {
+            return $this->errorMessage;
+        }
         $this->deleteModels();
         $this->dissociateRelations();
         $this->deleteRelations();
         $this->customDeleteActions();
-        $this->quotationRequest->delete();
+        if( count($this->errorMessage) === 0 ) {
+            $this->quotationRequest->delete();
+        }
 
         return $this->errorMessage;
     }
 
     /** Checks if the model can be deleted and sets error messages
      */
-    public function canDelete()
+    public function canDelete(): bool
     {
-        // Rewrite if to in_array check
-//        if(!in_array($this->quotationRequest->status_id, [2, 3, 4, 7, 8])) {
-//            if ($this->quotationRequest->status_id === 9) {
-//                array_push($this->errorMessage, "Een bezoek met status \"Afspraak gedaan\" kan niet verwijderd worden.");
-//            } else {
-//                array_push($this->errorMessage, "Er is nog een open offerteverzoek.");
-//            }
-//        }
-        //
-        if($this->quotationRequest->status->is_pending_status){
-            array_push($this->errorMessage, "Er is nog een " . ($this->quotationRequest->opportunityAction ? $this->quotationRequest->opportunityAction->name : "onbekend") . " met een onderhanden status " . ($this->quotationRequest->status ? $this->quotationRequest->status->name : "onbekend" ) );
-        }
+        // van hieruit altijd true
+        return true;
     }
 
     /** Deletes models recursive
